Separated Scene Graph from game logic; Scene must now be explicitly called to update game logic. Objects must now be initialized in scene instead. Scene must be derived (for manual memory management.).

+++ Engine/Animation/Private/Clip.cpp
@@ -82,7 +82,7 @@ void AnimSampler::Step(r32 gt)
 u32 AnimSampler::GetPaletteSz()
 {
   R_ASSERT(_pClip, "Clip is null for this sampler.");
-  return _pClip->_aAnimPoseSamples[_currPoseIdx]._aLocalPoses.size();
+  return (u32 )_pClip->_aAnimPoseSamples[_currPoseIdx]._aLocalPoses.size();
 }
